refactor deep link
